\name{ruinprob.finite.imps}
\alias{ruinprob.finite.imps}

\title{Approximation of the probability of ruin within a finite time horizon using importance sampling}

\description{
    This function calculates an approximation to the probability of ruin within a finite time
    horizon for a compound Poisson risk process that is perturbed by a Wiener process.  The
    approximation is based on importance sampling.
}

\usage{
    ruinprob.finite.imps()
}


%\details{
%}

\value{
    This function is not yet fully implemented.  At the moment it invisibly returns
    \code{NULL}.
}

\references{
Gatto, R. and Baumgartner, B. (to appear, in revision)  \emph{Saddlepoint
    approximations to the probability of ruin in finite time for the compound
    Poisson risk process perturbed by diffusion}. Methodology and Computing in
    Applied Probability \bold{TBA}(TBA), pp. 1--19.
}

%\note{
%%  ~~further notes~~
%}

%\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
%}

%\examples{
%}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
%\keyword{ ~kwd1 }
